refactor(startf): use stepfile to convert flat files to steps
