+++ transform/startf/testdata/fetch.star
@@ -1,5 +1,4 @@
 load("http.star", "http")
-load("qri.star", "qri")
 
 def download(ctx):
   res = http.get(test_server_url)
